refactor(store): extract lock machinery to store_lock.rs
The WS4 writer-probe additions pushed store.rs past the 1000-line module cap.
Move the advisory-flock behavior — writer_lock_is_held, WriterProbe /
probe_writer_lock, acquire_lock(_as), and the .lock.meta read/write helpers,
plus their tests — into a new src/store_lock.rs, mirroring the existing
store_blob.rs / store_gc.rs split. The lock TYPES (LockHolder / LockMeta) and
StoreError stay in store.rs; the entry points are re-exported from crate::store
so store_gc, main.rs, and callers are unaffected. store.rs: 871 lines.
